+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Changed
 
+### Changed
+
+  - Reset password and user invitation workflow enhancement
+
 ## [5.3.0] - 2020-10-09
 
 ### Added